+++ Runtime/Scripts/Extensions/Integrals/Int/IntExtensions.ToString.cs
@@ -9,7 +9,7 @@ namespace NumericMath
 	{
 		public static string ToBinaryString(this int value, int minLength = Int.BinaryLength)
 		{
-			return Convert.ToString(value, (int)Numeric.Base.Binary).PadLeft(minLength, Numeric.Zero);
+			return Convert.ToString(value, Numeric.Base.Binary).PadLeft(minLength, Numeric.Zero);
 		}
 
 		public static string ToHexString(this int value, int minLength = Int.HexLength)
